Linting header bannner components
